disable rpath
This has the advance that we dont get any lintian warnings but the drawback that afterwards
the drivers.so are no longer found. 

So this patch is DISABLED in series currently.

This is caused by the fact that dlopen does not consider the current dir if not told so 
(XnLinuxSharedLibs.cpp should be fixed for that).


Gbp-Pq: Name 0004-disable-rpath.patch